Build on Travis without shipping to web server.
No longer shipping artifacts to my own personal web server. Instead I'm
simply reporting the coverage to the command line. All I need to do is
indicate whether or not it is at 100%. I'm not going to want to go back
and audit it line by line.

Why do I need this when I'm submitting to Coveralls? Because Coveralls
only tracks line coverage, not branch coverage.